+++ bugtest_apps/qcolor_static_init_fiasco/other.cpp
@@ -1,0 +1,6 @@
+// other.cpp
+
+#include <QColor>
+#include "other.h"
+
+const QColor STATIC_NAME_INIT_OTHER_CPP("purple");  // may work, may fail