+++ bugtest_apps/qcolor_static_init_fiasco/other.h
@@ -1,0 +1,5 @@
+// other.h
+
+#pragma once
+
+extern const QColor STATIC_NAME_INIT_OTHER_CPP;