+++ server/tools/BUILD_REINSTALL_DEBIAN_PACKAGE.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+set -e
+
+TOOLDIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+
+echo "MAKING LINUX PACKAGES"
+${TOOLDIR}/MAKE_LINUX_PACKAGES.py
+echo "REINSTALLING DEBIAN PACKAGE"
+${TOOLDIR}/REINSTALL_DEBIAN_PACKAGE.sh